requirements.txt
blinker==1.9.0
cachelib==0.13.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
Flask==3.1.2
Flask-Caching==2.3.1
flask-cors==6.0.2
Flask-SQLAlchemy==3.1.1
fonttools==4.61.1
idna==3.11
iniconfig==2.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
nltk==3.9.2
numpy==2.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
pluggy==1.6.0
praw==7.8.1
prawcore==2.4.0
psycopg2==2.9.11
Pygments==2.19.2
pyparsing==3.3.1
pytest==9.0.2
pytest-mock==3.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
redis==7.1.0
regex==2025.11.3
requests==2.32.5
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.45
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.3
update-checker==0.18.0
urllib3==2.6.2
websocket-client==1.9.0
Werkzeug==3.1.0
wordcloud==1.9.5